+++ src/items.lua
@@ -1,12 +1,12 @@
 local _, Addon = ...
-local E = Addon:GetModule("Events")
+local E = Addon:GetModule("Events") ---@type Events
 local EquipmentSetsCache = Addon:GetModule("EquipmentSetsCache")
-local EventManager = Addon:GetModule("EventManager")
+local EventManager = Addon:GetModule("EventManager") ---@type EventManager
 local GetDetailedItemLevelInfo = C_Item.GetDetailedItemLevelInfo or GetDetailedItemLevelInfo
 local IsEquippableItem = C_Item.IsEquippableItem or IsEquippableItem
 local Items = Addon:GetModule("Items")
 local NUM_BAG_SLOTS = Addon.IS_RETAIL and NUM_TOTAL_EQUIPPED_BAG_SLOTS or NUM_BAG_SLOTS
-local TickerManager = Addon:GetModule("TickerManager")
+local TickerManager = Addon:GetModule("TickerManager") ---@type TickerManager
 
 -- Initialize cache table.
 Items.cache = {}
@@ -113,27 +113,18 @@ end
 -- Events
 -- ============================================================================
 
-do
-  local ticker
+-- Register events to trigger cache updates.
+EventManager:Once(E.Wow.PlayerLogin, function()
+  local debounce = TickerManager:NewDebouncer(0.01, updateCache)
+  debounce()
 
-  local function refreshTicker()
-    if ticker then
-      ticker:Restart()
-    else
-      ticker = TickerManager:NewTicker(0.01, updateCache, 1)
-    end
-  end
-
-  EventManager:Once(E.Wow.PlayerLogin, refreshTicker)
-
-  EventManager:On(E.Wow.BagUpdate, refreshTicker)
-  EventManager:On(E.Wow.BagUpdateDelayed, refreshTicker)
-  EventManager:On(E.Wow.EquipmentSetsChanged, refreshTicker)
-
+  EventManager:On(E.Wow.BagUpdate, debounce)
+  EventManager:On(E.Wow.BagUpdateDelayed, debounce)
+  EventManager:On(E.Wow.EquipmentSetsChanged, debounce)
   EventManager:On(E.BagsUpdated, function(allItemsCached)
-    if not allItemsCached then refreshTicker() end
+    if not allItemsCached then debounce() end
   end)
-end
+end)
 
 -- ============================================================================
 -- Bags
@@ -255,221 +246,229 @@ function Items:IsItemArtifactRelic(item)
 end
 
 -- ============================================================================
--- Initialize
+-- Suitable Items Table
 -- ============================================================================
 
--- Suitable items.
-EventManager:Once(E.Wow.PlayerLogin, function()
+--- Map of suitable armor and weapon subclasses based on the player's class.
+--- @type table<integer, table<integer, boolean>>
+Items.suitable = {
+  -- [classId] = { [subclassId] = boolean|nil }
+  [Enum.ItemClass.Armor] = {},
+  [Enum.ItemClass.Weapon] = {}
+}
+
+--- Updates the `Items.suitable` table.
+--- @param playerLevel integer
+local function updateSuitableTable(playerLevel)
+  local IS_LESSER_ARMOR_TYPE_SUITABLE = Addon.IS_VANILLA or (Addon.IS_CATA and playerLevel < 50)
   local _, class = UnitClass("player")
-  local suitable = {
-    -- [classId] = { [subclassId] = true }
-    [Enum.ItemClass.Armor] = {},
-    [Enum.ItemClass.Weapon] = {}
-  }
 
   -- Generic armor.
-  suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Generic] = true
-  suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Cosmetic] = true
+  Items.suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Generic] = true
+  Items.suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Cosmetic] = true
   -- Generic weapons.
-  suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Generic] = true
-  suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Fishingpole] = true
+  Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Generic] = true
+  Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Fishingpole] = true
 
   -- Warrior.
   if class == "WARRIOR" then
     -- Armor.
-    suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Cloth] = Addon.IS_VANILLA
-    suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Leather] = Addon.IS_VANILLA
-    suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Mail] = Addon.IS_VANILLA
-    suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Plate] = true
-    suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Shield] = true
+    Items.suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Cloth] = IS_LESSER_ARMOR_TYPE_SUITABLE
+    Items.suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Leather] = IS_LESSER_ARMOR_TYPE_SUITABLE
+    Items.suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Mail] = IS_LESSER_ARMOR_TYPE_SUITABLE
+    Items.suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Plate] = true
+    Items.suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Shield] = true
     -- Weapons.
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Axe1H] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Axe2H] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Bows] = Addon.IS_VANILLA or Addon.IS_CATA
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Crossbow] = Addon.IS_VANILLA or Addon.IS_CATA
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Dagger] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Guns] = Addon.IS_VANILLA or Addon.IS_CATA
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Mace1H] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Mace2H] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Polearm] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Staff] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Sword1H] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Sword2H] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Thrown] = Addon.IS_VANILLA or Addon.IS_CATA
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Unarmed] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Axe1H] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Axe2H] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Bows] = Addon.IS_VANILLA or Addon.IS_CATA
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Crossbow] = Addon.IS_VANILLA or Addon.IS_CATA
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Dagger] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Guns] = Addon.IS_VANILLA or Addon.IS_CATA
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Mace1H] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Mace2H] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Polearm] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Staff] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Sword1H] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Sword2H] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Thrown] = Addon.IS_VANILLA or Addon.IS_CATA
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Unarmed] = true
   end
 
   -- Paladin.
   if class == "PALADIN" then
     -- Armor.
-    suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Cloth] = Addon.IS_VANILLA
-    suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Leather] = Addon.IS_VANILLA
-    suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Libram] = Addon.IS_VANILLA
-    suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Mail] = Addon.IS_VANILLA
-    suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Plate] = true
-    suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Relic] = Addon.IS_CATA
-    suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Shield] = true
+    Items.suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Cloth] = IS_LESSER_ARMOR_TYPE_SUITABLE
+    Items.suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Leather] = IS_LESSER_ARMOR_TYPE_SUITABLE
+    Items.suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Libram] = Addon.IS_VANILLA
+    Items.suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Mail] = IS_LESSER_ARMOR_TYPE_SUITABLE
+    Items.suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Plate] = true
+    Items.suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Relic] = Addon.IS_CATA
+    Items.suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Shield] = true
     -- Weapons.
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Axe1H] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Axe2H] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Mace1H] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Mace2H] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Polearm] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Sword1H] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Sword2H] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Axe1H] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Axe2H] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Mace1H] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Mace2H] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Polearm] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Sword1H] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Sword2H] = true
   end
 
   -- Hunter.
   if class == "HUNTER" then
     -- Armor.
-    suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Cloth] = Addon.IS_VANILLA
-    suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Leather] = Addon.IS_VANILLA
-    suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Mail] = true
+    Items.suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Cloth] = IS_LESSER_ARMOR_TYPE_SUITABLE
+    Items.suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Leather] = IS_LESSER_ARMOR_TYPE_SUITABLE
+    Items.suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Mail] = true
     -- Weapons.
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Axe1H] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Axe2H] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Bows] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Crossbow] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Dagger] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Guns] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Polearm] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Staff] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Sword1H] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Sword2H] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Thrown] = Addon.IS_VANILLA or Addon.IS_CATA
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Unarmed] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Axe1H] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Axe2H] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Bows] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Crossbow] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Dagger] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Guns] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Polearm] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Staff] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Sword1H] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Sword2H] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Thrown] = Addon.IS_VANILLA or Addon.IS_CATA
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Unarmed] = true
   end
 
   -- Rogue.
   if class == "ROGUE" then
     -- Armor
-    suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Cloth] = Addon.IS_VANILLA
-    suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Leather] = true
+    Items.suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Cloth] = IS_LESSER_ARMOR_TYPE_SUITABLE
+    Items.suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Leather] = true
     -- Weapons
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Axe1H] = not Addon.IS_VANILLA
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Bows] = Addon.IS_VANILLA or Addon.IS_CATA
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Crossbow] = Addon.IS_VANILLA or Addon.IS_CATA
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Dagger] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Guns] = Addon.IS_VANILLA or Addon.IS_CATA
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Mace1H] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Sword1H] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Thrown] = Addon.IS_VANILLA or Addon.IS_CATA
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Unarmed] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Axe1H] = not Addon.IS_VANILLA
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Bows] = Addon.IS_VANILLA or Addon.IS_CATA
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Crossbow] = Addon.IS_VANILLA or Addon.IS_CATA
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Dagger] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Guns] = Addon.IS_VANILLA or Addon.IS_CATA
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Mace1H] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Sword1H] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Thrown] = Addon.IS_VANILLA or Addon.IS_CATA
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Unarmed] = true
   end
 
   -- Priest.
   if class == "PRIEST" then
     -- Armor
-    suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Cloth] = true
+    Items.suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Cloth] = true
     -- Weapons
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Dagger] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Mace1H] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Staff] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Wand] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Dagger] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Mace1H] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Staff] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Wand] = true
   end
 
   -- Death Knight.
   if class == "DEATHKNIGHT" then
     -- Armor.
-    suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Plate] = true
-    suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Relic] = Addon.IS_CATA
+    Items.suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Plate] = true
+    Items.suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Relic] = Addon.IS_CATA
     -- Weapons.
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Axe1H] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Axe2H] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Mace1H] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Mace2H] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Polearm] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Sword1H] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Sword2H] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Axe1H] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Axe2H] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Mace1H] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Mace2H] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Polearm] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Sword1H] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Sword2H] = true
   end
 
   -- Shaman.
   if class == "SHAMAN" then
     -- Armor.
-    suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Cloth] = Addon.IS_VANILLA
-    suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Leather] = Addon.IS_VANILLA
-    suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Mail] = true
-    suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Relic] = Addon.IS_CATA
-    suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Shield] = true
-    suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Totem] = Addon.IS_VANILLA
+    Items.suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Cloth] = IS_LESSER_ARMOR_TYPE_SUITABLE
+    Items.suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Leather] = IS_LESSER_ARMOR_TYPE_SUITABLE
+    Items.suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Mail] = true
+    Items.suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Relic] = Addon.IS_CATA
+    Items.suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Shield] = true
+    Items.suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Totem] = Addon.IS_VANILLA
     -- Weapons.
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Axe1H] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Axe2H] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Dagger] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Mace1H] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Mace2H] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Staff] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Unarmed] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Axe1H] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Axe2H] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Dagger] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Mace1H] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Mace2H] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Staff] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Unarmed] = true
   end
 
   -- Mage/Warlock.
   if class == "MAGE" or class == "WARLOCK" then
     -- Armor.
-    suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Cloth] = true
+    Items.suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Cloth] = true
     -- Weapons.
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Dagger] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Staff] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Sword1H] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Wand] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Dagger] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Staff] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Sword1H] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Wand] = true
   end
 
   -- Monk.
   if class == "MONK" then
     -- Armor.
-    suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Leather] = true
+    Items.suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Leather] = true
     -- Weapons.
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Axe1H] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Mace1H] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Polearm] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Staff] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Sword1H] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Unarmed] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Axe1H] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Mace1H] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Polearm] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Staff] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Sword1H] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Unarmed] = true
   end
 
   -- Druid.
   if class == "DRUID" then
     -- Armor.
-    suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Cloth] = Addon.IS_VANILLA
-    suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Idol] = Addon.IS_VANILLA
-    suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Leather] = true
-    suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Relic] = Addon.IS_CATA
+    Items.suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Cloth] = IS_LESSER_ARMOR_TYPE_SUITABLE
+    Items.suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Idol] = Addon.IS_VANILLA
+    Items.suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Leather] = true
+    Items.suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Relic] = Addon.IS_CATA
     -- Weapons.
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Dagger] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Mace1H] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Mace2H] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Staff] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Unarmed] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Bearclaw] = not Addon.IS_VANILLA
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Catclaw] = not Addon.IS_VANILLA
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Polearm] = not Addon.IS_VANILLA
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Dagger] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Mace1H] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Mace2H] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Staff] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Unarmed] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Bearclaw] = not Addon.IS_VANILLA
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Catclaw] = not Addon.IS_VANILLA
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Polearm] = not Addon.IS_VANILLA
   end
 
   -- Demon Hunter.
   if class == "DEMONHUNTER" then
     -- Armor.
-    suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Leather] = true
+    Items.suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Leather] = true
     -- Weapons.
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Axe1H] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Sword1H] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Unarmed] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Warglaive] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Axe1H] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Sword1H] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Unarmed] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Warglaive] = true
   end
 
   -- Evoker.
   if class == "EVOKER" then
     -- Armor.
-    suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Mail] = true
+    Items.suitable[Enum.ItemClass.Armor][Enum.ItemArmorSubclass.Mail] = true
     -- Weapons.
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Axe1H] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Axe2H] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Dagger] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Mace1H] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Mace2H] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Staff] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Sword1H] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Sword2H] = true
-    suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Unarmed] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Axe1H] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Axe2H] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Dagger] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Mace1H] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Mace2H] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Staff] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Sword1H] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Sword2H] = true
+    Items.suitable[Enum.ItemClass.Weapon][Enum.ItemWeaponSubclass.Unarmed] = true
   end
+end
 
-  Items.suitable = suitable
+EventManager:Once(E.Wow.PlayerLogin, function()
+  updateSuitableTable(UnitLevel("player"))
+  EventManager:On(E.Wow.PlayerLevelUp, updateSuitableTable)
 end)